+++ src/components/Album/style.css
@@ -13,7 +13,7 @@
 
 .album-art{
     overflow: hidden;
-    height: 200px;
+    height: 260px;
 }
 
 
@@ -23,7 +23,7 @@
 }
 
 .single-album .album-filter{
-    height: 200px;
+    height: 260px;
     width: 100%;
     position: absolute;
     z-index: 5;
